Remove implicit conversion from RTC_(D)CHECK_OPs
Implicit-conversion warnings are part of Chrome's (D)CHECK_OP macros so
in order to use those (and prevent webrtc roll breakages) we can't use
rtc::SafeEq etc, as they will implicitly sign convert parameters.

DO NOT SUBMIT

Bug: chromium:1216177

+++ modules/audio_coding/neteq/normal.cc
@@ -153,7 +153,7 @@ int Normal::Process(const int16_t* input,
   } else if (last_mode == NetEq::Mode::kRfc3389Cng) {
     RTC_DCHECK_EQ(output->Channels(), 1);  // Not adapted for multi-channel yet.
     static const size_t kCngLength = 48;
-    RTC_DCHECK_LE(8 * fs_mult, kCngLength);
+    RTC_DCHECK_LE(static_cast<size_t>(8 * fs_mult), kCngLength);
     int16_t cng_output[kCngLength];
     ComfortNoiseDecoder* cng_decoder = decoder_database_->GetActiveCngDecoder();
